Publish only if tag

+++ travis-build.sh
@@ -12,11 +12,11 @@ if [[ -n $TRAVIS_TAG ]] || [[ $TRAVIS_BRANCH == 'master' && $TRAVIS_PULL_REQUEST
 
   echo "Publishing archives"
 
-  # if [[ -n $TRAVIS_TAG ]]; then
-  #    ./gradlew bintrayUpload || EXIT_STATUS=$?
+  if [[ -n $TRAVIS_TAG ]]; then
+      ./gradlew bintrayUpload || EXIT_STATUS=$?
   #else
   #    ./gradlew publish || EXIT_STATUS=$?
-  #fi
+  fi
 
   ./gradlew docs || EXIT_STATUS=$?
 
@@ -37,13 +37,13 @@ if [[ -n $TRAVIS_TAG ]] || [[ $TRAVIS_BRANCH == 'master' && $TRAVIS_PULL_REQUEST
   fi
 
   # If there is a tag present then this becomes the latest
-  #  if [[ -n $TRAVIS_TAG ]]; then
+  if [[ -n $TRAVIS_TAG ]]; then
         git rm -rf latest/
         mkdir -p latest
         cp -r ../build/docs/. ./latest/
         git add latest/*
 
-        version="3.0.1"
+        version="$TRAVIS_TAG"
         version=${version:1}
         majorVersion=${version:0:4}
         majorVersion="${majorVersion}x"
@@ -56,7 +56,7 @@ if [[ -n $TRAVIS_TAG ]] || [[ $TRAVIS_BRANCH == 'master' && $TRAVIS_PULL_REQUEST
         cp -r ../build/docs/. "./$majorVersion/"
         git add "$majorVersion/*"
 
-  #  fi
+  fi
 
     git commit -a -m "Updating docs for Travis build: https://travis-ci.org/$TRAVIS_REPO_SLUG/builds/$TRAVIS_BUILD_ID"
     git push origin HEAD
